requirements.txt
accelerate==1.10.0
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cmake==4.1.0
contourpy==1.3.0
cycler==0.12.1
deepspeed==0.17.4
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.18.0
fonttools==4.59.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.7
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.1
networkx==3.2.1
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.65
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.7.0.72
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.33
requests==2.32.4
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.35.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tifffile==2024.8.30
tokenizers==0.21.4
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
transformers==4.55.2
triton==2.2.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
zipp==3.23.0
